🦄 refactor: 优化类型定义
