+++ electron/main/ipc/ipc-tray.ts
@@ -1,3 +1,4 @@
+import { PlayModePayload } from "@shared";
 import { ipcMain } from "electron";
 import { getMainTray } from "../tray";
 import { appName } from "../utils/config";
@@ -27,7 +28,7 @@ const initTrayIpc = (): void => {
   });
 
   // 播放模式切换
-  ipcMain.on("play-mode-change", (_, data: { repeatMode: any; shuffleMode: any }) => {
+  ipcMain.on("play-mode-change", (_, data: PlayModePayload) => {
     tray?.setPlayMode(data.repeatMode, data.shuffleMode);
   });
